Wczytywanie przeciwników z pliku mapy

+++ SDL/Player.h
@@ -7,16 +7,16 @@
 class Player {
 
 public:
-	Player(const char* texturesheet, float x, float y);
+	Player(const char* texturesheet, int x, int y);
 	~Player();
 
 	void Update();
 	void Render();
 
-	float getPositionX();
-	float getPositionY();
-	void setPositionX(float xpos);
-	void setPositionY(float ypos);
+	int getPositionX();
+	int getPositionY();
+	void setPositionX(int xpos);
+	void setPositionY(int ypos);
 
 	void setMapSize(int size_x, int size_y);
 	void setCollidingTiles(std::vector<class Tile*> tiles);
@@ -25,8 +25,8 @@ public:
 
 private:
 	int oSize;
-	float xpos;
-	float ypos;
+	int xpos;
+	int ypos;
 	
 	// kolizja
 	int map_size_x;
@@ -38,8 +38,8 @@ private:
 	bool movLeft;
 	bool movRight;
 
-	float velocityFactorX;
-	float velocityFactorY;
+	int velocityFactorX;
+	int velocityFactorY;
 
 	Velocity* oVelocity;
 	SDL_Texture* oTexture;
